refactor(ml_engine): extract metrics/ subpackage (PR-2 of decomp)
Moves metrics.py (1018 LOC) and real_world_benchmarks.py (1378 LOC)
out of the flat services/ directory into a focused metrics/
subpackage:

  metrics.py                -> metrics/compute.py
  real_world_benchmarks.py  -> metrics/real_world_benchmarks.py

Adds metrics/__init__.py re-exporting MetricsService, VintageAnalyser,
ks_statistic, psi, psi_by_feature, brier_decomposition, and
RealWorldBenchmarks so the existing `from apps.ml_engine.services.metrics
import MetricsService` style continues to work unchanged.

Updates the 2 source files that import RealWorldBenchmarks directly
(generate_data management command + loan_performance_simulator). All
other metrics.* imports work via the re-export.

DEFERRED from the spec: the in-file split of metrics.py into focused
modules (compute / fairness / calibration / ranking). That split is
genuinely risky line-by-line work that benefits from its own atomic
commit + test cycle. This PR ships the subpackage organisation; the
split lands as a follow-up PR if scheduled.

Pure refactor — file moves + 2 import-path updates + new __init__.py.
Zero behaviour change; full backend suite passes 1613/1613 (34
pre-existing skips, mainly hypothesis-missing tests).

Implements PR-2 of docs/superpowers/specs/2026-05-25-ml-engine-decomposition-design.md.
Stacks on PR-1 (#196). Base targets feat/ml-engine-decomp-pr1-external;
retarget to master after #196 merges.

Recon caveat from PR-1 applied: grep broadened to relative imports +
full backend suite ran once before declaring done. Zero stragglers
found, zero unrelated frontend files swept up.
